chore: conformance review of transaction types & field mappings (#33)
- Replace rmpv::Value passthrough with fully typed structs for heartbeat
  (HeartbeatTxnFields, HeartbeatProof), state proof (StateProofBody + 8
  nested types), and state proof message (StateProofMessage)
- Add missing Access field (al) with ResourceRef, HoldingRef, LocalsRef structs
- Fix type widths: AssetParams.decimals u64→u32, Transaction.extra_program_pages u64→u32
- Add canonical encode functions for all 15 new typed structs
- Confirm LogicSig.lmsig field exists in Go reference (no change needed)
- 810 tests pass, zero failures

+++ crates/core/algo-ledger/src/apply.rs
@@ -903,7 +903,7 @@ fn apply_appl<L: crate::store_trait::LedgerStore>(
 
         let global_schema = txn.global_state_schema.clone().unwrap_or_default();
         let local_schema = txn.local_state_schema.clone().unwrap_or_default();
-        let extra_pages = txn.extra_program_pages as u32;
+        let extra_pages = txn.extra_program_pages;
 
         store.set_app_params(
             app_id,
